Split code in to multiple files, add docstrings
