+++ terminal_test.go
@@ -17,7 +17,9 @@ var TestFiles = []string{
 func loadFixture(base string, ext string) []byte {
 	filename := fmt.Sprintf("fixtures/%s.%s", base, ext)
 	data, err := ioutil.ReadFile(filename)
-	check(err)
+	if err != nil {
+		fmt.Errorf("could not load fixture %s: %v", filename, err)
+	}
 	return data
 }
